+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates `iconColorsArray` to include `ui.warning.tertiary`.
-- Utilizies onBlur events to clear error state in comment and email input elements in `FeedbackBox` component.
+- Utilizies onChange events to clear error state in comment and email input elements in `FeedbackBox` component.
 
 ## 4.1.6 (April 2, 2026)